int changed to float

+++ src/main/java/com/example/springdocker/model/MyMathCalc.java
@@ -11,7 +11,7 @@ public int add(int a,int b){
 public int multiply (int a,int b){
         return a*b;
 }
-public int divide(int a,int b){
+public float divide(int a,int b){
         return (a/b);
 }
 }
